wgpu

+++ brewlib/Graphics/Renderers/IQuadRenderer.cs
@@ -1,10 +1,9 @@
 ﻿namespace BrewLib.Graphics.Renderers;
 
-using System;
 using System.Numerics;
 using System.Runtime.InteropServices;
-using BrewLib.Graphics.Textures;
 using SixLabors.ImageSharp.PixelFormats;
+using Textures;
 
 public interface IQuadRenderer : IPrimitiveRenderer
 {
@@ -15,18 +14,18 @@ public interface IQuadRenderer : IPrimitiveRenderer
 struct QuadPrimitive
 {
     public Vector2 vec1;
-    public Half u1, v1;
+    public float u1, v1;
     public Rgba32 color1;
 
     public Vector2 vec2;
-    public Half u2, v2;
+    public float u2, v2;
     public Rgba32 color2;
 
     public Vector2 vec3;
-    public Half u3, v3;
+    public float u3, v3;
     public Rgba32 color3;
 
     public Vector2 vec4;
-    public Half u4, v4;
+    public float u4, v4;
     public Rgba32 color4;
 }
